Refactor xtask bench
